Change JSON encoding flag from JSON_ENSURE_ASCII to JSON_COMPACT.
JSON_COMPACT results in a, as the name suggests, more compact JSON
representation allowing to send more characters as part of a push
notification. Further, not using JSON_ENSURE_ASCII is necessary
for languages that do not use the latin alphabet (Arabic, CJK, etc),
otherwise the character limit for notifications is ridiculously
short.